+++ extensions/curl/CHANGELOG.md
@@ -1,5 +1,10 @@
 # cURL Changelog
 
+## [Fix: Windows Shortcuts] - 2025-09-18
+
+- Updated `@raycast/api` to the latest version
+- Added cross-platform shortcut support for Windows
+
 ## [Feat: Add Windows Support] - 2025-09-09
 
 - Added support for Windows